Move scheduler bits around a little
