+++ letsencrypt-win-simple/Plugin/IISSiteServerPlugin.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Security.Cryptography.X509Certificates;
 using LetsEncrypt.ACME.Simple.Configuration;
 using LetsEncrypt.ACME.Simple.Extensions;
+using LetsEncrypt.ACME.Simple.Schedules;
 using Serilog;
 
 namespace LetsEncrypt.ACME.Simple
